fix: Mypage 비밀번호찾기/회원탈퇴/프로필이미지변경 (#141)
* feat: MobileSideBar/stores/로그인 로직구현

* fix: 로그인 로직구현(store에 로그인여부 저장, local storage에 액세스토큰저장, App에 로그인여부 state 연결)

* fix: 질문작성페이지 axios 통신로직에  Headers 설정

* fix: 상세페이지 store생성, 연결/ 페이지라우트params변경

* fix: 마크다운뷰어, 에디터 데이터  가져오기 완료

* fix: 세부기능 구현/SideBar 연결

* fix: RightSideBar 구조 수정

* fix: 상세페이지 삭제한 코드 추가/axios controller/MyPage기능/Nav logout기능

* fix: Mypage 비밀번호찾기/회원탈퇴/프로필이미지변경

+++ client/src/Components/MyPageEdit.js
@@ -45,17 +45,22 @@ const MyPageEdit = ({ parsed, setRender }) => {
 
   const handleDeleteProfile = (e) => {
     e.preventDefault();
-    axios
-      .delete(`/members/${parsed.memberId}`, {
-        headers: {
-          Authorization: 'Bearer ' + localStorage.getItem('ACCESS_TOKEN'),
-        },
-      })
-      .then(() => {
-        localStorage.clear();
-        navigate('/');
-      })
-      .catch((err) => alert(err.response.data.message));
+    if (window.confirm('확인을 누르면 회원 정보가 삭제됩니다.')) {
+      axios
+        .delete(`/members/${parsed.memberId}`, {
+          headers: {
+            Authorization: 'Bearer ' + localStorage.getItem('ACCESS_TOKEN'),
+          },
+        })
+        .then(() => {
+          localStorage.clear();
+          alert('그동안 이용해주셔서 감사합니다.');
+          navigate('/');
+        })
+        .catch((err) => alert(err.response.data.message));
+    } else {
+      return;
+    }
   };
 
   return (
